requirements.txt
bcrypt===4.2.0
cffi===1.16.0
cryptography===42.0.8
decorator===5.1.1
Deprecated===1.2.14
invoke===2.2.0
paramiko===3.4.0
psycopg2===2.9.9
pycparser===2.22
PyNaCl===1.5.0
PyYAML===6.0.1
sshtunnel===0.4.0
wrapt===1.16.